Better command output + refactor tests
