+++ resources/views/terminal/fail.blade.php
@@ -1,0 +1,1 @@
+<div class="text-red-500 mx-2 my-1">{{$value}}</div>
