+++ resources/views/terminal/success.blade.php
@@ -1,0 +1,1 @@
+<div class="text-green-500 mx-2 my-1">{{$value}}</div>
