+++ src/Console/ListCommand.php
@@ -5,13 +5,13 @@ namespace Nutgram\Laravel\Console;
 use Closure;
 use Illuminate\Console\Command;
 use Illuminate\Support\Arr;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 use ReflectionProperty;
 use SergiX44\Nutgram\Handlers\Handler;
+use SergiX44\Nutgram\Handlers\Type\Command as NutgramCommand;
 use SergiX44\Nutgram\Nutgram;
-use SergiX44\Nutgram\Telegram\Attributes\MessageTypes;
-use SergiX44\Nutgram\Telegram\Properties\MessageType;
-use Symfony\Component\Console\Terminal;
+use function Termwind\render;
 
 class ListCommand extends Command
 {
@@ -19,83 +19,44 @@ class ListCommand extends Command
 
     protected $description = 'List all registered handlers';
 
-    public function handle(): int
+    public function handle(Nutgram $bot): int
     {
-        $bot = app(Nutgram::class);
-
-        $refHandlers = new ReflectionProperty(Nutgram::class, "handlers");
-        $refHandlers->setAccessible(true);
-
-        $handlers = collect(Arr::dot($refHandlers->getValue($bot)))
-            ->map(function (Handler $handler, string $key) {
-                $refCallable = new ReflectionProperty(Handler::class, "callable");
-                $refCallable->setAccessible(true);
-                $callable = $refCallable->getValue($handler);
-
-                return [
-                    'handler' => $this->getHandlerName($key),
-                    'arg' => $handler->getPattern(),
-                    'subtype' => $this->getMessageSubtype($key),
-                    'callable' => $this->getCallableName($callable),
-                ];
-            });
+        $handlers = $this->getHandlers($bot)->map(fn (Handler $handler, string $key) => [
+            'handler' => $this->getHandlerName($key, $handler instanceof NutgramCommand),
+            'pattern' => $handler->getPattern(),
+            'callable' => $this->getCallableName($handler),
+        ]);
 
         if ($handlers->isEmpty()) {
             $this->warn("No handlers have been registered.");
             return 0;
         }
 
-        $terminalWidth = (new Terminal)->getWidth();
-        $padding = 2;
-        $maxHandler = (int)$handlers->max(fn ($item) => strlen($item['handler'])) + 1;
-
-        foreach ($handlers as $data) {
-            [$handler, $arg, $subtype, $callable] = array_values($data);
-
-            $handler = str_pad($handler ?? '', $maxHandler);
-            $arg = (!in_array($handler, ['onText', 'onCommand']) ? $arg : $subtype);
-            $arg = $arg ?: $subtype;
-
-            $dots = str_repeat('.', max($terminalWidth - mb_strlen($handler . $arg . $callable) - ($padding * 2), 0));
-
-            $output = "<fg=yellow>$handler</>";
-            $output .= "<fg=gray>$arg</>";
-            $output .= "<fg=blue>$dots</>";
-            $output .= "<fg=blue>$callable</>";
-
-            $output = str_repeat(' ', $padding) . $output . str_repeat(' ', $padding);
-
-            $this->line($output);
-        }
+        render(view('terminal::list', ['items' => $handlers]));
 
         return 0;
     }
 
-    protected function getHandlerName(string $signature): ?string
+    protected function getHandlers(Nutgram $bot): Collection
     {
-        $signature = Str::lower($signature);
+        $refHandlers = new ReflectionProperty(Nutgram::class, "handlers");
+        $refHandlers->setAccessible(true);
+        return collect(Arr::dot($refHandlers->getValue($bot)));
+    }
 
+    protected function getHandlerName(string $signature, bool $isCommand = false): string
+    {
+        if ($isCommand) {
+            return 'onCommand';
+        }
+
+        return $this->resolveUpdateListener(Str::lower($signature));
+    }
+
+    protected function resolveUpdateListener(string $signature): string
+    {
         return match (Str::before($signature, '.')) {
-            'message' => value(
-                function () use ($signature) {
-                    $subHandlerSignature = Str::after($signature, 'message.');
-                    [$subHandlerName, $subHandlerKey] = array_pad(explode('.', $subHandlerSignature), 2, null);
-
-                    if ($subHandlerName === 'text') {
-                        if (Str::startsWith($subHandlerKey ?? '', '/')) {
-                            return 'onCommand';
-                        }
-
-                        return 'onText';
-                    }
-
-                    if (MessageType::tryFrom($subHandlerName)===null) {
-                        return 'onMessage';
-                    }
-
-                    return 'onMessageType';
-                }
-            ),
+            'message' => $this->resolveMessageListener($signature),
             'edited_message' => 'onEditedMessage',
             'channel_post' => 'onChannelPost',
             'edited_channel_post' => 'onEditedChannelPost',
@@ -104,37 +65,80 @@ class ListCommand extends Command
             'callback_query' => 'onCallbackQuery',
             'shipping_query' => 'onShippingQuery',
             'pre_checkout_query' => 'onPreCheckoutQuery',
-            'poll' => 'onPoll',
+            'poll' => 'onUpdatePoll',
             'poll_answer' => 'onPollAnswer',
             'my_chat_member' => 'onMyChatMember',
             'chat_member' => 'onChatMember',
             'chat_join_request' => 'onChatJoinRequest',
             'api_error' => 'onApiError',
             'exception' => 'onException',
-            default => null,
+            default => 'unknown',
         };
     }
 
-    protected function getMessageSubtype(string $value): ?string
+    protected function resolveMessageListener(string $signature): string
     {
-        $value = Str::lower($value);
+        $subHandlerSignature = Str::after($signature, 'message.');
+        $subHandlerName = explode('.', $subHandlerSignature)[0] ?? null;
 
-        if (!Str::startsWith($value, 'message.')) {
-            return null;
+        if(is_numeric($subHandlerName)) {
+            return 'onMessage';
         }
 
-        $subHandlerSignature = Str::after($value, 'message.');
-        [$subHandlerName] = array_pad(explode('.', $subHandlerSignature), 1, null);
-
-        if (MessageType::tryFrom($subHandlerName)===null) {
-            return null;
-        }
-
-        return sprintf("MessageType::%s", Str::upper($subHandlerName ?? 'UNKNOWN'));
+        return match($subHandlerName){
+            'text' => 'onText',
+            'animation' => 'onAnimation',
+            'audio' => 'onAudio',
+            'document' => 'onDocument',
+            'photo' => 'onPhoto',
+            'sticker' => 'onSticker',
+            'video' => 'onVideo',
+            'video_note' => 'onVideoNote',
+            'voice' => 'onVoice',
+            'contact' => 'onContact',
+            'dice' => 'onDice',
+            'game' => 'onGame',
+            'poll' => 'onMessagePoll',
+            'venue' => 'onVenue',
+            'location' => 'onLocation',
+            'new_chat_members' => 'onNewChatMembers',
+            'left_chat_member' => 'onLeftChatMember',
+            'new_chat_title' => 'onNewChatTitle',
+            'new_chat_photo' => 'onNewChatPhoto',
+            'delete_chat_photo' => 'onDeleteChatPhoto',
+            'group_chat_created' => 'onGroupChatCreated',
+            'supergroup_chat_created' => 'onSupergroupChatCreated',
+            'channel_chat_created' => 'onChannelChatCreated',
+            'message_auto_delete_timer_changed' => 'onMessageAutoDeleteTimerChanged',
+            'migrate_to_chat_id' => 'onMigrateToChatId',
+            'migrate_from_chat_id' => 'onMigrateFromChatId',
+            'pinned_message' => 'onPinnedMessage',
+            'invoice' => 'onInvoice',
+            'successful_payment' => 'onSuccessfulPayment',
+            'connected_website' => 'onConnectedWebsite',
+            'passport_data' => 'onPassportData',
+            'proximity_alert_triggered' => 'onProximityAlertTriggered',
+            'forum_topic_created' => 'onForumTopicCreated',
+            'forum_topic_edited' => 'onForumTopicEdited',
+            'forum_topic_closed' => 'onForumTopicClosed',
+            'forum_topic_reopened' => 'onForumTopicReopened',
+            'video_chat_scheduled' => 'onVideoChatScheduled',
+            'video_chat_started' => 'onVideoChatStarted',
+            'video_chat_ended' => 'onVideoChatEnded',
+            'video_chat_participants_invited' => 'onVideoChatParticipantsInvited',
+            'web_app_data' => 'onWebAppData',
+            default => 'unknown',
+        };
     }
 
-    protected function getCallableName(mixed $callable): string
+    protected function getCallableName(Handler $handler): string
     {
+        // get callable value
+        $refCallable = new ReflectionProperty(Handler::class, "callable");
+        $refCallable->setAccessible(true);
+        $callable = $refCallable->getValue($handler);
+
+        // parse callable
         if (is_string($callable)) {
             return trim($callable);
         }
