+++ tests/Commands/HookRemoveTest.php
@@ -1,0 +1,28 @@
+<?php
+
+use Mockery\MockInterface;
+use Nutgram\Laravel\Console\HookRemoveCommand;
+use SergiX44\Nutgram\Nutgram;
+
+test('nutgram:hook:remove removes the bot webhook', function () {
+    $this->mock(Nutgram::class, function (MockInterface $mock) {
+        $mock->shouldReceive('deleteWebhook')->with(false)->andReturn(0);
+    });
+
+    $this->artisan(HookRemoveCommand::class, ['--drop-pending-updates' => false])
+        ->expectsOutput('Bot webhook removed.')
+        ->assertSuccessful();
+});
+
+test('nutgram:hook:remove removes the bot webhook and the pending updates', function () {
+    $this->mock(Nutgram::class, function (MockInterface $mock) {
+        $mock->shouldReceive('deleteWebhook')->with([
+            'drop_pending_updates' => true,
+        ])->andReturn(0);
+    });
+
+    $this->artisan(HookRemoveCommand::class, ['--drop-pending-updates' => true])
+        ->expectsOutput('Pending updates dropped.')
+        ->expectsOutput('Bot webhook removed.')
+        ->assertSuccessful();
+});
